requirements.txt
aws-glue-schema-registry==1.1.3
boto3==1.34.56
botocore==1.34.56
confluent-kafka==2.3.0
Faker==24.0.0
fastavro==1.9.4
fastjsonschema==2.19.1
jmespath==1.0.1
kafka-python==2.0.2
orjson==3.6.9
pretty==0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.0
six==1.16.0
urllib3==2.0.7
